+++ src/util/dynamicArray.h
@@ -10,15 +10,15 @@
 template <typename T, typename S = void>
 class DynamicArray {
 	public:
-		size_t head;
-		size_t size;
+		uint64_t head;
+		uint64_t size;
 
 		DynamicArray() {
 			this->array = nullptr;
 			this->dontDelete = true;
 		}
 
-		DynamicArray(size_t size) {
+		DynamicArray(uint64_t size) {
 			this->parent = nullptr;
 			this->init = nullptr;
 			this->onRealloc = nullptr;
@@ -31,7 +31,7 @@ class DynamicArray {
 
 		DynamicArray(
 			S* parent,
-			size_t size,
+			uint64_t size,
 			void (*init) (S* parent, T* location),
 			void (*onRealloc) (S* parent)
 		) {
@@ -74,7 +74,7 @@ class DynamicArray {
 				this->array = array;
 
 				if(this->init != nullptr) {
-					for(size_t i = this->size; i < this->size * 2; i++) {
+					for(uint64_t i = this->size; i < this->size * 2; i++) {
 						(*this->init)(this->parent, &this->array[i]);
 					}
 				}
@@ -104,7 +104,7 @@ class DynamicArray {
 			this->array = array;
 
 			if(this->init != nullptr) {
-				for(size_t i = 0; i < this->size; i++) {
+				for(uint64_t i = 0; i < this->size; i++) {
 					(*this->init)(this->parent, &(this->array[i]));
 				}
 			}
